wang-landau/pyproject.toml
[project]
name = "wang-landau"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Arvin Kushwaha", email = "arvin.singh.kushwaha@gmail.com" }
]
dependencies = [
    "jaxtyping>=0.2.34",
    "jax[cuda12]>=0.4.33",
    "polars>=1.7.1",
    "matplotlib-backend-sixel>=0.1.0",
    "jaxlib[cuda12]>=0.4.33",
]
readme = "README.md"
requires-python = ">= 3.12"
license = { text = "MIT" }

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint.isort]
lines-between-types = 1
relative-imports-order = "closest-to-furthest"

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.basedpyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownLambdaType = false
reportUntypedFunctionDecorator = false
